requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.4.0
numpy==1.26.4